requirements.txt
langchain
langserve
langchain_community
langchain_core
python-dotenv
streamlit
fastapi
uvicorn
sse_starlette
bs4
beautifulsoup4
playwright
tiktoken
scrapegraphai
nest_asyncio
pypdf 
chroma
chromadb
faiss-cpu
sentence-transformers